feat: add configurable event size and checkout interval for rrweb events

+++ server/store/sessions.go
@@ -260,7 +260,7 @@ func (s *Store) GetSessionChunk(sessionID string, seq int) ([]json.RawMessage, e
 		return nil, err
 	}
 	defer zr.Close()
-	raw, err := io.ReadAll(io.LimitReader(zr, IngestBodyLimit))
+	raw, err := io.ReadAll(io.LimitReader(zr, int64(IngestBodyLimit)))
 	if err != nil {
 		return nil, err
 	}
